requirements.txt
aiofiles==24.1.0
asyncpg==0.30.0
autopep8==2.3.1
bcrypt==4.2.1
black==24.10.0
blinker==1.9.0
click==8.1.7
Flask==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
priority==2.0.0
pycodestyle==2.12.1
PyJWT==2.10.0
python-dotenv==1.0.1
Quart==0.19.8
Werkzeug==3.1.3
wsproto==1.2.0
